pyproject.toml
[project]
name = "restaurant-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "python-dotenv",
]
